Added isAdmin middleware to protect admin-only routes

+++ routes/api.php
@@ -16,6 +16,7 @@ use App\Http\Controllers\UserFavoritesController;
 use App\Http\Controllers\UserNotificationsController;
 use App\Http\Controllers\UserSettingsController;
 use App\Http\Middleware\CheckAuth;
+use App\Http\Middleware\IsAdmin;
 use Illuminate\Support\Facades\Route;
 
 Route::prefix('v1')->name('api.')
@@ -68,33 +69,33 @@ Route::prefix('v1')->name('api.')
         Route::prefix('chats')->name('chats.')->controller(ChatController::class)
             ->middleware(CheckAuth::class)
             ->group(function () {
-                Route::get('/', 'index')->name('index'); // TODO: админ онли
+                Route::get('/', 'index')->name('index')->middleware(IsAdmin::class);
                 Route::post('/', 'store')->name('store');
                 Route::get('{chat}', 'show')->name('show');
                 Route::match(['put', 'patch'], '{chat}', 'update')->name('update');
                 Route::delete('{chat}', 'destroy')->name('destroy');
 
-                Route::get('{chat}/messages', 'messages')->name('messages'); // TODO: админ онли
+                Route::get('{chat}/messages', 'messages')->name('messages')->middleware(IsAdmin::class);
             });
 
         Route::prefix('covers')->name('covers.')->controller(CoverController::class)
             ->middleware(CheckAuth::class)
             ->group(function () {
-                Route::get('/', 'index')->name('index'); // TODO: админ онли
+                Route::get('/', 'index')->name('index')->middleware(IsAdmin::class);
                 Route::post('/', 'store')->name('store');
                 Route::get('{cover}', 'show')->name('show')->withoutMiddleware(CheckAuth::class);
-                Route::match(['put', 'patch'], '{cover}', 'update')->name('update'); // TODO: админ онли
+                Route::match(['put', 'patch'], '{cover}', 'update')->name('update')->middleware(IsAdmin::class);
                 Route::delete('{cover}', 'destroy')->name('destroy');
             });
 
         Route::prefix('deals')->name('deals.')->controller(DealController::class)
             ->middleware(CheckAuth::class)
             ->group(function () {
-                Route::get('/', 'index')->name('index'); // TODO: админ онли
+                Route::get('/', 'index')->name('index')->middleware(IsAdmin::class);
                 Route::post('/', 'store')->name('store');
                 Route::get('{deal}', 'show')->name('show');
-                Route::match(['put', 'patch'], '{deal}', 'update')->name('update'); // TODO: админ онли
-                Route::delete('{deal}', 'destroy')->name('destroy'); // TODO: админ онли
+                Route::match(['put', 'patch'], '{deal}', 'update')->name('update')->middleware(IsAdmin::class);
+                Route::delete('{deal}', 'destroy')->name('destroy')->middleware(IsAdmin::class);
             });
 
         Route::prefix('genres')->name('genres.')->controller(GenreController::class)
@@ -110,10 +111,10 @@ Route::prefix('v1')->name('api.')
         Route::prefix('photos')->name('photos.')->controller(PhotoController::class)
             ->middleware(CheckAuth::class)
             ->group(function () {
-                Route::get('/', 'index')->name('index'); // TODO: админ онли
+                Route::get('/', 'index')->name('index')->middleware(IsAdmin::class);
                 Route::post('/', 'store')->name('store');
                 Route::get('{photo}', 'show')->name('show')->withoutMiddleware(CheckAuth::class);
-                Route::match(['put', 'patch'], '{photo}', 'update')->name('update'); // TODO: админ онли
+                Route::match(['put', 'patch'], '{photo}', 'update')->name('update')->middleware(IsAdmin::class);
                 Route::delete('{photo}', 'destroy')->name('destroy');
             });
 
@@ -121,16 +122,16 @@ Route::prefix('v1')->name('api.')
             ->middleware(CheckAuth::class)
             ->group(function () {
                 Route::get('/', 'index')->name('index');
-                Route::post('/', 'store')->name('store'); // TODO: админ онли
+                Route::post('/', 'store')->name('store')->middleware(IsAdmin::class);
                 Route::get('{setting}', 'show')->name('show');
-                Route::match(['put', 'patch'], '{setting}', 'update')->name('update'); // TODO: админ онли
-                Route::delete('{setting}', 'destroy')->name('destroy'); // TODO: админ онли
+                Route::match(['put', 'patch'], '{setting}', 'update')->name('update')->middleware(IsAdmin::class);
+                Route::delete('{setting}', 'destroy')->name('destroy')->middleware(IsAdmin::class);
             });
 
         Route::prefix('users')->name('users.')->controller(UserController::class)
             ->middleware(CheckAuth::class)
             ->group(function () {
-                Route::get('/', 'index')->name('index'); // TODO: админ онли
+                Route::get('/', 'index')->name('index')->middleware(IsAdmin::class);
                 Route::post('/', 'store')->name('store');
                 Route::get('{user}', 'show')->name('show')->withoutMiddleware(CheckAuth::class);
                 Route::match(['put', 'patch'], '{user}', 'update')->name('update');
